+++ SurvivorMS/Gui.cpp
@@ -134,7 +134,7 @@ gui::DropDownList::DropDownList(
 		sf::Color(255, 255, 255, 200), sf::Color(255, 255, 255, 255), sf::Color(20, 20, 20, 50)
 	);
 
-	for (size_t i = 0; i < elementsCount; i++) {
+	for (unsigned i = 0; i < elementsCount; i++) {
 		this->list.push_back(
 			new gui::Button(
 				x, y + ((i+1) * height),
@@ -233,12 +233,20 @@ gui::TextureSelector::TextureSelector(float x, float y, float width, float heigh
 
 	if (this->sheet.getGlobalBounds().width > this->bounds.getGlobalBounds().width) {
 
-		this->sheet.setTextureRect(sf::IntRect(0, 0, this->bounds.getGlobalBounds().width, this->bounds.getGlobalBounds().height));
+		this->sheet.setTextureRect(sf::IntRect(
+			0, 0, 
+			static_cast<int>(this->bounds.getGlobalBounds().width), 
+			static_cast<int>(this->bounds.getGlobalBounds().height)
+		));
 	}
 	
 	if (this->sheet.getGlobalBounds().height > this->bounds.getGlobalBounds().height) {
 		
-		this->sheet.setTextureRect(sf::IntRect(0, 0, this->bounds.getGlobalBounds().height, this->bounds.getGlobalBounds().width));
+		this->sheet.setTextureRect(sf::IntRect(
+			0, 0, 
+			static_cast<int>(this->bounds.getGlobalBounds().height),
+				static_cast<int>(this->bounds.getGlobalBounds().width)
+		));
 
 	}
